app/pyproject.toml
[tool.poetry]
name = "payslip-extractor"
version = "0.1.0"
description = "Continuously extract payslip data from iCloud email and upload to Google Sheets"
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
imapclient = "^2.3.0"
pyzmail36 = "^1.0.3"
PyPDF2 = "^3.0.1"
google-auth = "^2.22.0"
google-auth-oauthlib = "^1.0.0"
google-api-python-client = "^2.88.0"
gspread = "^5.7.2"
regex = "^2023.8.8"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

